moved logic into comparePurchaseUrgency
