+++ src/views/List.jsx
@@ -4,110 +4,14 @@ import { ListItem } from '../components';
 
 // ADDED
 import { getDaysBetweenDates } from '../utils';
+import { comparePurchaseUrgency } from '../api/firebase';
 
 export function List({ data, listToken }) {
 	const navigate = useNavigate();
 	const [searchTerm, setSearchTerm] = useState('');
 	const [copied, setCopied] = useState('');
 
-	let currentTime = new Date().getTime(); //in MS
-	// separating out the logic for active/inactive
-	const checkIfActive = (refDate, current) => {
-		return getDaysBetweenDates(refDate, current) < 60;
-	};
-	/*
-	 * not currently split out comparePurchaseUrgency
-	 * lots of duplication in ListItem code (maybe pass more of the daysNext? or keep the daysNext/daysSince extraction from other try)
-	 * useEffect prob not the best implementation for real-time update of this kind
-	 */
-
-	const sortedOverdueList = data
-		.filter((item) => {
-			let refDate = item.dateLastPurchased
-				? item.dateLastPurchased.toMillis()
-				: item.dateCreated.toMillis();
-
-			// for overdues:
-			let daysToNext = getDaysBetweenDates(
-				currentTime,
-				item.dateNextPurchased.toMillis(),
-			);
-			return checkIfActive(refDate, currentTime) && daysToNext < 0;
-		})
-		.sort(
-			(a, b) => b.dateNextPurchased.toMillis() - a.dateNextPurchased.toMillis(),
-		) //descending order here - farther out = MORE overdue
-		// alphabetize same daysToNext items:
-		.sort((a, b) => {
-			//get days to next
-			let aDaysToNext = getDaysBetweenDates(
-				currentTime,
-				a.dateNextPurchased.toMillis(),
-			);
-			let bDaysToNext = getDaysBetweenDates(
-				currentTime,
-				b.dateNextPurchased.toMillis(),
-			);
-			return aDaysToNext - bDaysToNext || a.name.localeCompare(b.name);
-		});
-
-	const sortedActiveListNoOverdues = data
-		.filter((item) => {
-			let refDate = item.dateLastPurchased
-				? item.dateLastPurchased.toMillis()
-				: item.dateCreated.toMillis();
-			// for overdues: omit here
-			let daysToNext = getDaysBetweenDates(
-				currentTime,
-				item.dateNextPurchased.toMillis(),
-			);
-			return checkIfActive(refDate, currentTime) && daysToNext >= 0;
-		})
-		.sort(
-			(a, b) => a.dateNextPurchased.toMillis() - b.dateNextPurchased.toMillis(),
-		) //ascending order - proximity to currentTime
-		.sort((a, b) => {
-			//get days to next
-			let aDaysToNext = getDaysBetweenDates(
-				currentTime,
-				a.dateNextPurchased.toMillis(),
-			);
-			let bDaysToNext = getDaysBetweenDates(
-				currentTime,
-				b.dateNextPurchased.toMillis(),
-			);
-			return aDaysToNext - bDaysToNext || a.name.localeCompare(b.name);
-			// if we wanted reverse alph order: return - ( a.id - b.id  ||  a.name.localeCompare(b.name) )
-			// does not function: return bDaysToNext - DaysToNext || [a.name,b.name].sort()
-		});
-
-	const sortedInactiveListByDateNext = data
-		.filter((item) => {
-			let refDate = item.dateLastPurchased
-				? item.dateLastPurchased.toMillis()
-				: item.dateCreated.toMillis();
-			return !checkIfActive(refDate, currentTime);
-		})
-		.sort(
-			(a, b) => a.dateNextPurchased.toMillis() - b.dateNextPurchased.toMillis(),
-		)
-		.sort((a, b) => {
-			//get days to next
-			let aDaysToNext = getDaysBetweenDates(
-				currentTime,
-				a.dateNextPurchased.toMillis(),
-			);
-			let bDaysToNext = getDaysBetweenDates(
-				currentTime,
-				b.dateNextPurchased.toMillis(),
-			);
-			return aDaysToNext - bDaysToNext || a.name.localeCompare(b.name);
-		});
-
-	const sortedFullList = sortedOverdueList.concat(
-		sortedActiveListNoOverdues,
-		sortedInactiveListByDateNext,
-	);
+	const sortedFullList = comparePurchaseUrgency(data);
 
 	useEffect(() => {
 		if (copied) setTimeout(() => setCopied(''), 2000);
